+++ CHANGELOG.md
@@ -289,3 +289,4 @@
 - 2024-02-12: Bench descend, tz country, leads, mod, +2
 - 2024-02-12: Integ campaign, users, qualify, follow up, +6
 - 2024-02-13: Pool lib, mod, frontier, chat, +4
+- 2024-02-14: Log conf, leads, lib, bettercontact, +3